feat: implement policies and permissions

+++ app/Http/Controllers/ImportInvoice/InvoiceController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\ImportInvoice;
 
+use App\Constants\Abilities;
 use App\Contracts\PaymentGatewayContract;
 use App\Exports\InvoiceTemplateExport;
 use App\Http\Controllers\Controller;
@@ -10,6 +11,7 @@ use App\Imports\InvoicesImport;
 use App\Infrastructure\Persistence\Models\Invoice;
 use App\ViewModels\Invoice\InvoiceIndexViewModel;
 use App\ViewModels\Invoice\InvoiceMicrositeIndexViewModel;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Artisan;
@@ -22,6 +24,7 @@ use Symfony\Component\HttpFoundation\BinaryFileResponse;
 
 class InvoiceController extends Controller
 {
+    use AuthorizesRequests;
     public function downloadTemplate(): BinaryFileResponse
     {
         Log::info('Generating the invoice template...');
@@ -40,6 +43,8 @@ class InvoiceController extends Controller
 
     public function index(Request $request, InvoiceIndexViewModel $viewModel, InvoiceMicrositeIndexViewModel $viewModelMicrosite): Response
     {
+        $this->authorize(Abilities::VIEW_ANY->value, Invoice::class);
+
         Artisan::call('app:transactions-consult');
         $search = $request->only(['search']);
 
@@ -61,6 +66,8 @@ class InvoiceController extends Controller
 
     public function show(string $id): Response
     {
+        $this->authorize(Abilities::VIEW->value, Invoice::class);
+
         $invoice = Invoice::with('microsite')->findOrFail($id);
 
         return Inertia::render('Invoices/Show', compact('invoice'));
@@ -68,6 +75,7 @@ class InvoiceController extends Controller
 
     public function destroy(string $id): RedirectResponse
     {
+        $this->authorize(Abilities::DELETE->value, Invoice::class);
         Invoice::findOrFail($id)->delete();
 
         return to_route('invoices.index')->with('success', 'Invoice deleted successfully.');
